Refactored Collectible, Build ItemGenerator
Refactored Collectibles, so that they are now also an "Interface", so i can add it to Objects as i generate them.
Now you can generate Ingame-Items out of code. The item needs a predefined model (for the collision and the sprite), a value and an item_name. Then the ItemGenerator can create the Item ingame, so it has all the Skills I create for them (Collectable, Item etc.)