Addition of Processors
I added 2 Processors  (Damage and Stat) with the purpose of pipelining the process of managing the different sources of Values for the different parts. Those are the places where i define how damage is calculated and define how and where i get certain values from progress values. For example: To calculate the strength the character has, the StatProcessor requests the info from other processors or Information-containers. Like Strength:
First the stats from the CharacterStats, then from the Equipment, Perks etc.

This way i can also call methods that are dependent on the attacker and the victim. For example if the attacker has a bonus damage against a certain monstertype. Or an item gives a damage bonus to a certain monster type.

This makes it easier for me to implement new Progress-Systems and just use it in one place, so i can keep the rest of the code clean and expand it easier